feat(NET-1180): split WorkerAssignment/PortalAssignment into their own schema files
Addresses define-null's PR review comment -- the new root types lived inline in
assignment.fbs instead of getting their own worker_assignment.fbs/portal_assignment.fbs.

worker_assignment.fbs and portal_assignment.fbs now `include "assignment.fbs"` to reuse
WorkerStatus/WorkerId/EncryptedHeaders/WorkerEntry, each generating their own
*_generated.rs (flatc emits one file per schema, cross-referencing shared types via a
hardcoded `use crate::<name>_generated::*`). assignment_fb.rs is now a thin facade
re-exporting all three generated modules into one flat namespace, so builder.rs/reader.rs
keep referencing `assignment_fb::Foo` unchanged -- purely a schema/codegen
reorganization, no wire format or public API change.

Also: narrowing #![allow(clippy::all, ...)] to just the generated-code modules (it was
previously blanket-suppressing lints across all of assignment_fb.rs, including
hand-written code) surfaced one genuine pre-existing nit -- a redundant `.clone()` on
`PeerId` (which is `Copy`) in a test -- now fixed.
