+++ crates/assignments/schema/gen/worker_assignment_generated.rs
@@ -1,0 +1,572 @@
+// automatically generated by the FlatBuffers compiler, do not modify
+
+
+// @generated
+
+use crate::assignment_generated::*;
+use core::mem;
+use core::cmp::Ordering;
+
+extern crate flatbuffers;
+use self::flatbuffers::{EndianScalar, Follow};
+
+pub enum WorkerAssignmentOffset {}
+#[derive(Copy, Clone, PartialEq)]
+
+pub struct WorkerAssignment<'a> {
+  pub _tab: flatbuffers::Table<'a>,
+}
+
+impl<'a> flatbuffers::Follow<'a> for WorkerAssignment<'a> {
+  type Inner = WorkerAssignment<'a>;
+  #[inline]
+  unsafe fn follow(buf: &'a [u8], loc: usize) -> Self::Inner {
+    Self { _tab: flatbuffers::Table::new(buf, loc) }
+  }
+}
+
+impl<'a> WorkerAssignment<'a> {
+  pub const VT_DATASETS: flatbuffers::VOffsetT = 4;
+  pub const VT_WORKERS: flatbuffers::VOffsetT = 6;
+
+  #[inline]
+  pub unsafe fn init_from_table(table: flatbuffers::Table<'a>) -> Self {
+    WorkerAssignment { _tab: table }
+  }
+  #[allow(unused_mut)]
+  pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr, A: flatbuffers::Allocator + 'bldr>(
+    _fbb: &'mut_bldr mut flatbuffers::FlatBufferBuilder<'bldr, A>,
+    args: &'args WorkerAssignmentArgs<'args>
+  ) -> flatbuffers::WIPOffset<WorkerAssignment<'bldr>> {
+    let mut builder = WorkerAssignmentBuilder::new(_fbb);
+    if let Some(x) = args.workers { builder.add_workers(x); }
+    if let Some(x) = args.datasets { builder.add_datasets(x); }
+    builder.finish()
+  }
+
+
+  #[inline]
+  pub fn datasets(&self) -> flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WorkerAssignmentDataset<'a>>> {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WorkerAssignmentDataset>>>>(WorkerAssignment::VT_DATASETS, None).unwrap()}
+  }
+  #[inline]
+  pub fn workers(&self) -> flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WorkerEntry<'a>>> {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WorkerEntry>>>>(WorkerAssignment::VT_WORKERS, None).unwrap()}
+  }
+}
+
+impl flatbuffers::Verifiable for WorkerAssignment<'_> {
+  #[inline]
+  fn run_verifier(
+    v: &mut flatbuffers::Verifier, pos: usize
+  ) -> Result<(), flatbuffers::InvalidFlatbuffer> {
+    use self::flatbuffers::Verifiable;
+    v.visit_table(pos)?
+     .visit_field::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'_, flatbuffers::ForwardsUOffset<WorkerAssignmentDataset>>>>("datasets", Self::VT_DATASETS, true)?
+     .visit_field::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'_, flatbuffers::ForwardsUOffset<WorkerEntry>>>>("workers", Self::VT_WORKERS, true)?
+     .finish();
+    Ok(())
+  }
+}
+pub struct WorkerAssignmentArgs<'a> {
+    pub datasets: Option<flatbuffers::WIPOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WorkerAssignmentDataset<'a>>>>>,
+    pub workers: Option<flatbuffers::WIPOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WorkerEntry<'a>>>>>,
+}
+impl<'a> Default for WorkerAssignmentArgs<'a> {
+  #[inline]
+  fn default() -> Self {
+    WorkerAssignmentArgs {
+      datasets: None, // required field
+      workers: None, // required field
+    }
+  }
+}
+
+pub struct WorkerAssignmentBuilder<'a: 'b, 'b, A: flatbuffers::Allocator + 'a> {
+  fbb_: &'b mut flatbuffers::FlatBufferBuilder<'a, A>,
+  start_: flatbuffers::WIPOffset<flatbuffers::TableUnfinishedWIPOffset>,
+}
+impl<'a: 'b, 'b, A: flatbuffers::Allocator + 'a> WorkerAssignmentBuilder<'a, 'b, A> {
+  #[inline]
+  pub fn add_datasets(&mut self, datasets: flatbuffers::WIPOffset<flatbuffers::Vector<'b , flatbuffers::ForwardsUOffset<WorkerAssignmentDataset<'b >>>>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WorkerAssignment::VT_DATASETS, datasets);
+  }
+  #[inline]
+  pub fn add_workers(&mut self, workers: flatbuffers::WIPOffset<flatbuffers::Vector<'b , flatbuffers::ForwardsUOffset<WorkerEntry<'b >>>>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WorkerAssignment::VT_WORKERS, workers);
+  }
+  #[inline]
+  pub fn new(_fbb: &'b mut flatbuffers::FlatBufferBuilder<'a, A>) -> WorkerAssignmentBuilder<'a, 'b, A> {
+    let start = _fbb.start_table();
+    WorkerAssignmentBuilder {
+      fbb_: _fbb,
+      start_: start,
+    }
+  }
+  #[inline]
+  pub fn finish(self) -> flatbuffers::WIPOffset<WorkerAssignment<'a>> {
+    let o = self.fbb_.end_table(self.start_);
+    self.fbb_.required(o, WorkerAssignment::VT_DATASETS,"datasets");
+    self.fbb_.required(o, WorkerAssignment::VT_WORKERS,"workers");
+    flatbuffers::WIPOffset::new(o.value())
+  }
+}
+
+impl core::fmt::Debug for WorkerAssignment<'_> {
+  fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+    let mut ds = f.debug_struct("WorkerAssignment");
+      ds.field("datasets", &self.datasets());
+      ds.field("workers", &self.workers());
+      ds.finish()
+  }
+}
+pub enum WorkerAssignmentDatasetOffset {}
+#[derive(Copy, Clone, PartialEq)]
+
+pub struct WorkerAssignmentDataset<'a> {
+  pub _tab: flatbuffers::Table<'a>,
+}
+
+impl<'a> flatbuffers::Follow<'a> for WorkerAssignmentDataset<'a> {
+  type Inner = WorkerAssignmentDataset<'a>;
+  #[inline]
+  unsafe fn follow(buf: &'a [u8], loc: usize) -> Self::Inner {
+    Self { _tab: flatbuffers::Table::new(buf, loc) }
+  }
+}
+
+impl<'a> WorkerAssignmentDataset<'a> {
+  pub const VT_ID: flatbuffers::VOffsetT = 4;
+  pub const VT_CHUNKS: flatbuffers::VOffsetT = 6;
+  pub const VT_LAST_BLOCK: flatbuffers::VOffsetT = 8;
+
+  #[inline]
+  pub unsafe fn init_from_table(table: flatbuffers::Table<'a>) -> Self {
+    WorkerAssignmentDataset { _tab: table }
+  }
+  #[allow(unused_mut)]
+  pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr, A: flatbuffers::Allocator + 'bldr>(
+    _fbb: &'mut_bldr mut flatbuffers::FlatBufferBuilder<'bldr, A>,
+    args: &'args WorkerAssignmentDatasetArgs<'args>
+  ) -> flatbuffers::WIPOffset<WorkerAssignmentDataset<'bldr>> {
+    let mut builder = WorkerAssignmentDatasetBuilder::new(_fbb);
+    builder.add_last_block(args.last_block);
+    if let Some(x) = args.chunks { builder.add_chunks(x); }
+    if let Some(x) = args.id { builder.add_id(x); }
+    builder.finish()
+  }
+
+
+  #[inline]
+  pub fn id(&self) -> &'a str {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WorkerAssignmentDataset::VT_ID, None).unwrap()}
+  }
+  #[inline]
+  pub fn key_compare_less_than(&self, o: &WorkerAssignmentDataset) -> bool {
+    self.id() < o.id()
+  }
+
+  #[inline]
+  pub fn key_compare_with_value(&self, val: & str) -> ::core::cmp::Ordering {
+    let key = self.id();
+    key.cmp(val)
+  }
+  #[inline]
+  pub fn chunks(&self) -> flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WorkerAssignmentChunk<'a>>> {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WorkerAssignmentChunk>>>>(WorkerAssignmentDataset::VT_CHUNKS, None).unwrap()}
+  }
+  #[inline]
+  pub fn last_block(&self) -> u64 {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<u64>(WorkerAssignmentDataset::VT_LAST_BLOCK, Some(0)).unwrap()}
+  }
+}
+
+impl flatbuffers::Verifiable for WorkerAssignmentDataset<'_> {
+  #[inline]
+  fn run_verifier(
+    v: &mut flatbuffers::Verifier, pos: usize
+  ) -> Result<(), flatbuffers::InvalidFlatbuffer> {
+    use self::flatbuffers::Verifiable;
+    v.visit_table(pos)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("id", Self::VT_ID, true)?
+     .visit_field::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'_, flatbuffers::ForwardsUOffset<WorkerAssignmentChunk>>>>("chunks", Self::VT_CHUNKS, true)?
+     .visit_field::<u64>("last_block", Self::VT_LAST_BLOCK, false)?
+     .finish();
+    Ok(())
+  }
+}
+pub struct WorkerAssignmentDatasetArgs<'a> {
+    pub id: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub chunks: Option<flatbuffers::WIPOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<WorkerAssignmentChunk<'a>>>>>,
+    pub last_block: u64,
+}
+impl<'a> Default for WorkerAssignmentDatasetArgs<'a> {
+  #[inline]
+  fn default() -> Self {
+    WorkerAssignmentDatasetArgs {
+      id: None, // required field
+      chunks: None, // required field
+      last_block: 0,
+    }
+  }
+}
+
+pub struct WorkerAssignmentDatasetBuilder<'a: 'b, 'b, A: flatbuffers::Allocator + 'a> {
+  fbb_: &'b mut flatbuffers::FlatBufferBuilder<'a, A>,
+  start_: flatbuffers::WIPOffset<flatbuffers::TableUnfinishedWIPOffset>,
+}
+impl<'a: 'b, 'b, A: flatbuffers::Allocator + 'a> WorkerAssignmentDatasetBuilder<'a, 'b, A> {
+  #[inline]
+  pub fn add_id(&mut self, id: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WorkerAssignmentDataset::VT_ID, id);
+  }
+  #[inline]
+  pub fn add_chunks(&mut self, chunks: flatbuffers::WIPOffset<flatbuffers::Vector<'b , flatbuffers::ForwardsUOffset<WorkerAssignmentChunk<'b >>>>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WorkerAssignmentDataset::VT_CHUNKS, chunks);
+  }
+  #[inline]
+  pub fn add_last_block(&mut self, last_block: u64) {
+    self.fbb_.push_slot::<u64>(WorkerAssignmentDataset::VT_LAST_BLOCK, last_block, 0);
+  }
+  #[inline]
+  pub fn new(_fbb: &'b mut flatbuffers::FlatBufferBuilder<'a, A>) -> WorkerAssignmentDatasetBuilder<'a, 'b, A> {
+    let start = _fbb.start_table();
+    WorkerAssignmentDatasetBuilder {
+      fbb_: _fbb,
+      start_: start,
+    }
+  }
+  #[inline]
+  pub fn finish(self) -> flatbuffers::WIPOffset<WorkerAssignmentDataset<'a>> {
+    let o = self.fbb_.end_table(self.start_);
+    self.fbb_.required(o, WorkerAssignmentDataset::VT_ID,"id");
+    self.fbb_.required(o, WorkerAssignmentDataset::VT_CHUNKS,"chunks");
+    flatbuffers::WIPOffset::new(o.value())
+  }
+}
+
+impl core::fmt::Debug for WorkerAssignmentDataset<'_> {
+  fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+    let mut ds = f.debug_struct("WorkerAssignmentDataset");
+      ds.field("id", &self.id());
+      ds.field("chunks", &self.chunks());
+      ds.field("last_block", &self.last_block());
+      ds.finish()
+  }
+}
+pub enum WorkerAssignmentChunkOffset {}
+#[derive(Copy, Clone, PartialEq)]
+
+pub struct WorkerAssignmentChunk<'a> {
+  pub _tab: flatbuffers::Table<'a>,
+}
+
+impl<'a> flatbuffers::Follow<'a> for WorkerAssignmentChunk<'a> {
+  type Inner = WorkerAssignmentChunk<'a>;
+  #[inline]
+  unsafe fn follow(buf: &'a [u8], loc: usize) -> Self::Inner {
+    Self { _tab: flatbuffers::Table::new(buf, loc) }
+  }
+}
+
+impl<'a> WorkerAssignmentChunk<'a> {
+  pub const VT_FIRST_BLOCK: flatbuffers::VOffsetT = 4;
+  pub const VT_ID: flatbuffers::VOffsetT = 6;
+  pub const VT_DATASET_ID: flatbuffers::VOffsetT = 8;
+  pub const VT_SIZE: flatbuffers::VOffsetT = 10;
+  pub const VT_DATASET_BASE_URL: flatbuffers::VOffsetT = 12;
+  pub const VT_SCHEMA_ID: flatbuffers::VOffsetT = 14;
+  pub const VT_TABLES_PRESENT: flatbuffers::VOffsetT = 16;
+  pub const VT_WORKER_INDEXES: flatbuffers::VOffsetT = 18;
+
+  #[inline]
+  pub unsafe fn init_from_table(table: flatbuffers::Table<'a>) -> Self {
+    WorkerAssignmentChunk { _tab: table }
+  }
+  #[allow(unused_mut)]
+  pub fn create<'bldr: 'args, 'args: 'mut_bldr, 'mut_bldr, A: flatbuffers::Allocator + 'bldr>(
+    _fbb: &'mut_bldr mut flatbuffers::FlatBufferBuilder<'bldr, A>,
+    args: &'args WorkerAssignmentChunkArgs<'args>
+  ) -> flatbuffers::WIPOffset<WorkerAssignmentChunk<'bldr>> {
+    let mut builder = WorkerAssignmentChunkBuilder::new(_fbb);
+    builder.add_first_block(args.first_block);
+    if let Some(x) = args.worker_indexes { builder.add_worker_indexes(x); }
+    if let Some(x) = args.tables_present { builder.add_tables_present(x); }
+    builder.add_schema_id(args.schema_id);
+    if let Some(x) = args.dataset_base_url { builder.add_dataset_base_url(x); }
+    builder.add_size(args.size);
+    if let Some(x) = args.dataset_id { builder.add_dataset_id(x); }
+    if let Some(x) = args.id { builder.add_id(x); }
+    builder.finish()
+  }
+
+
+  #[inline]
+  pub fn first_block(&self) -> u64 {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<u64>(WorkerAssignmentChunk::VT_FIRST_BLOCK, Some(0)).unwrap()}
+  }
+  #[inline]
+  pub fn key_compare_less_than(&self, o: &WorkerAssignmentChunk) -> bool {
+    self.first_block() < o.first_block()
+  }
+
+  #[inline]
+  pub fn key_compare_with_value(&self, val: u64) -> ::core::cmp::Ordering {
+    let key = self.first_block();
+    key.cmp(&val)
+  }
+  #[inline]
+  pub fn id(&self) -> &'a str {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WorkerAssignmentChunk::VT_ID, None).unwrap()}
+  }
+  #[inline]
+  pub fn dataset_id(&self) -> &'a str {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WorkerAssignmentChunk::VT_DATASET_ID, None).unwrap()}
+  }
+  #[inline]
+  pub fn size(&self) -> u32 {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<u32>(WorkerAssignmentChunk::VT_SIZE, Some(0)).unwrap()}
+  }
+  #[inline]
+  pub fn dataset_base_url(&self) -> &'a str {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<flatbuffers::ForwardsUOffset<&str>>(WorkerAssignmentChunk::VT_DATASET_BASE_URL, None).unwrap()}
+  }
+  #[inline]
+  pub fn schema_id(&self) -> i32 {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<i32>(WorkerAssignmentChunk::VT_SCHEMA_ID, Some(0)).unwrap()}
+  }
+  #[inline]
+  pub fn tables_present(&self) -> Option<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<&'a str>>> {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<&'a str>>>>(WorkerAssignmentChunk::VT_TABLES_PRESENT, None)}
+  }
+  #[inline]
+  pub fn worker_indexes(&self) -> flatbuffers::Vector<'a, u16> {
+    // Safety:
+    // Created from valid Table for this object
+    // which contains a valid value in this slot
+    unsafe { self._tab.get::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'a, u16>>>(WorkerAssignmentChunk::VT_WORKER_INDEXES, None).unwrap()}
+  }
+}
+
+impl flatbuffers::Verifiable for WorkerAssignmentChunk<'_> {
+  #[inline]
+  fn run_verifier(
+    v: &mut flatbuffers::Verifier, pos: usize
+  ) -> Result<(), flatbuffers::InvalidFlatbuffer> {
+    use self::flatbuffers::Verifiable;
+    v.visit_table(pos)?
+     .visit_field::<u64>("first_block", Self::VT_FIRST_BLOCK, false)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("id", Self::VT_ID, true)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("dataset_id", Self::VT_DATASET_ID, true)?
+     .visit_field::<u32>("size", Self::VT_SIZE, false)?
+     .visit_field::<flatbuffers::ForwardsUOffset<&str>>("dataset_base_url", Self::VT_DATASET_BASE_URL, true)?
+     .visit_field::<i32>("schema_id", Self::VT_SCHEMA_ID, false)?
+     .visit_field::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'_, flatbuffers::ForwardsUOffset<&'_ str>>>>("tables_present", Self::VT_TABLES_PRESENT, false)?
+     .visit_field::<flatbuffers::ForwardsUOffset<flatbuffers::Vector<'_, u16>>>("worker_indexes", Self::VT_WORKER_INDEXES, true)?
+     .finish();
+    Ok(())
+  }
+}
+pub struct WorkerAssignmentChunkArgs<'a> {
+    pub first_block: u64,
+    pub id: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub dataset_id: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub size: u32,
+    pub dataset_base_url: Option<flatbuffers::WIPOffset<&'a str>>,
+    pub schema_id: i32,
+    pub tables_present: Option<flatbuffers::WIPOffset<flatbuffers::Vector<'a, flatbuffers::ForwardsUOffset<&'a str>>>>,
+    pub worker_indexes: Option<flatbuffers::WIPOffset<flatbuffers::Vector<'a, u16>>>,
+}
+impl<'a> Default for WorkerAssignmentChunkArgs<'a> {
+  #[inline]
+  fn default() -> Self {
+    WorkerAssignmentChunkArgs {
+      first_block: 0,
+      id: None, // required field
+      dataset_id: None, // required field
+      size: 0,
+      dataset_base_url: None, // required field
+      schema_id: 0,
+      tables_present: None,
+      worker_indexes: None, // required field
+    }
+  }
+}
+
+pub struct WorkerAssignmentChunkBuilder<'a: 'b, 'b, A: flatbuffers::Allocator + 'a> {
+  fbb_: &'b mut flatbuffers::FlatBufferBuilder<'a, A>,
+  start_: flatbuffers::WIPOffset<flatbuffers::TableUnfinishedWIPOffset>,
+}
+impl<'a: 'b, 'b, A: flatbuffers::Allocator + 'a> WorkerAssignmentChunkBuilder<'a, 'b, A> {
+  #[inline]
+  pub fn add_first_block(&mut self, first_block: u64) {
+    self.fbb_.push_slot::<u64>(WorkerAssignmentChunk::VT_FIRST_BLOCK, first_block, 0);
+  }
+  #[inline]
+  pub fn add_id(&mut self, id: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WorkerAssignmentChunk::VT_ID, id);
+  }
+  #[inline]
+  pub fn add_dataset_id(&mut self, dataset_id: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WorkerAssignmentChunk::VT_DATASET_ID, dataset_id);
+  }
+  #[inline]
+  pub fn add_size(&mut self, size: u32) {
+    self.fbb_.push_slot::<u32>(WorkerAssignmentChunk::VT_SIZE, size, 0);
+  }
+  #[inline]
+  pub fn add_dataset_base_url(&mut self, dataset_base_url: flatbuffers::WIPOffset<&'b  str>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WorkerAssignmentChunk::VT_DATASET_BASE_URL, dataset_base_url);
+  }
+  #[inline]
+  pub fn add_schema_id(&mut self, schema_id: i32) {
+    self.fbb_.push_slot::<i32>(WorkerAssignmentChunk::VT_SCHEMA_ID, schema_id, 0);
+  }
+  #[inline]
+  pub fn add_tables_present(&mut self, tables_present: flatbuffers::WIPOffset<flatbuffers::Vector<'b , flatbuffers::ForwardsUOffset<&'b  str>>>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WorkerAssignmentChunk::VT_TABLES_PRESENT, tables_present);
+  }
+  #[inline]
+  pub fn add_worker_indexes(&mut self, worker_indexes: flatbuffers::WIPOffset<flatbuffers::Vector<'b , u16>>) {
+    self.fbb_.push_slot_always::<flatbuffers::WIPOffset<_>>(WorkerAssignmentChunk::VT_WORKER_INDEXES, worker_indexes);
+  }
+  #[inline]
+  pub fn new(_fbb: &'b mut flatbuffers::FlatBufferBuilder<'a, A>) -> WorkerAssignmentChunkBuilder<'a, 'b, A> {
+    let start = _fbb.start_table();
+    WorkerAssignmentChunkBuilder {
+      fbb_: _fbb,
+      start_: start,
+    }
+  }
+  #[inline]
+  pub fn finish(self) -> flatbuffers::WIPOffset<WorkerAssignmentChunk<'a>> {
+    let o = self.fbb_.end_table(self.start_);
+    self.fbb_.required(o, WorkerAssignmentChunk::VT_ID,"id");
+    self.fbb_.required(o, WorkerAssignmentChunk::VT_DATASET_ID,"dataset_id");
+    self.fbb_.required(o, WorkerAssignmentChunk::VT_DATASET_BASE_URL,"dataset_base_url");
+    self.fbb_.required(o, WorkerAssignmentChunk::VT_WORKER_INDEXES,"worker_indexes");
+    flatbuffers::WIPOffset::new(o.value())
+  }
+}
+
+impl core::fmt::Debug for WorkerAssignmentChunk<'_> {
+  fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+    let mut ds = f.debug_struct("WorkerAssignmentChunk");
+      ds.field("first_block", &self.first_block());
+      ds.field("id", &self.id());
+      ds.field("dataset_id", &self.dataset_id());
+      ds.field("size", &self.size());
+      ds.field("dataset_base_url", &self.dataset_base_url());
+      ds.field("schema_id", &self.schema_id());
+      ds.field("tables_present", &self.tables_present());
+      ds.field("worker_indexes", &self.worker_indexes());
+      ds.finish()
+  }
+}
+#[inline]
+/// Verifies that a buffer of bytes contains a `WorkerAssignment`
+/// and returns it.
+/// Note that verification is still experimental and may not
+/// catch every error, or be maximally performant. For the
+/// previous, unchecked, behavior use
+/// `root_as_worker_assignment_unchecked`.
+pub fn root_as_worker_assignment(buf: &[u8]) -> Result<WorkerAssignment, flatbuffers::InvalidFlatbuffer> {
+  flatbuffers::root::<WorkerAssignment>(buf)
+}
+#[inline]
+/// Verifies that a buffer of bytes contains a size prefixed
+/// `WorkerAssignment` and returns it.
+/// Note that verification is still experimental and may not
+/// catch every error, or be maximally performant. For the
+/// previous, unchecked, behavior use
+/// `size_prefixed_root_as_worker_assignment_unchecked`.
+pub fn size_prefixed_root_as_worker_assignment(buf: &[u8]) -> Result<WorkerAssignment, flatbuffers::InvalidFlatbuffer> {
+  flatbuffers::size_prefixed_root::<WorkerAssignment>(buf)
+}
+#[inline]
+/// Verifies, with the given options, that a buffer of bytes
+/// contains a `WorkerAssignment` and returns it.
+/// Note that verification is still experimental and may not
+/// catch every error, or be maximally performant. For the
+/// previous, unchecked, behavior use
+/// `root_as_worker_assignment_unchecked`.
+pub fn root_as_worker_assignment_with_opts<'b, 'o>(
+  opts: &'o flatbuffers::VerifierOptions,
+  buf: &'b [u8],
+) -> Result<WorkerAssignment<'b>, flatbuffers::InvalidFlatbuffer> {
+  flatbuffers::root_with_opts::<WorkerAssignment<'b>>(opts, buf)
+}
+#[inline]
+/// Verifies, with the given verifier options, that a buffer of
+/// bytes contains a size prefixed `WorkerAssignment` and returns
+/// it. Note that verification is still experimental and may not
+/// catch every error, or be maximally performant. For the
+/// previous, unchecked, behavior use
+/// `root_as_worker_assignment_unchecked`.
+pub fn size_prefixed_root_as_worker_assignment_with_opts<'b, 'o>(
+  opts: &'o flatbuffers::VerifierOptions,
+  buf: &'b [u8],
+) -> Result<WorkerAssignment<'b>, flatbuffers::InvalidFlatbuffer> {
+  flatbuffers::size_prefixed_root_with_opts::<WorkerAssignment<'b>>(opts, buf)
+}
+#[inline]
+/// Assumes, without verification, that a buffer of bytes contains a WorkerAssignment and returns it.
+/// # Safety
+/// Callers must trust the given bytes do indeed contain a valid `WorkerAssignment`.
+pub unsafe fn root_as_worker_assignment_unchecked(buf: &[u8]) -> WorkerAssignment {
+  flatbuffers::root_unchecked::<WorkerAssignment>(buf)
+}
+#[inline]
+/// Assumes, without verification, that a buffer of bytes contains a size prefixed WorkerAssignment and returns it.
+/// # Safety
+/// Callers must trust the given bytes do indeed contain a valid size prefixed `WorkerAssignment`.
+pub unsafe fn size_prefixed_root_as_worker_assignment_unchecked(buf: &[u8]) -> WorkerAssignment {
+  flatbuffers::size_prefixed_root_unchecked::<WorkerAssignment>(buf)
+}
+#[inline]
+pub fn finish_worker_assignment_buffer<'a, 'b, A: flatbuffers::Allocator + 'a>(
+    fbb: &'b mut flatbuffers::FlatBufferBuilder<'a, A>,
+    root: flatbuffers::WIPOffset<WorkerAssignment<'a>>) {
+  fbb.finish(root, None);
+}
+
+#[inline]
+pub fn finish_size_prefixed_worker_assignment_buffer<'a, 'b, A: flatbuffers::Allocator + 'a>(fbb: &'b mut flatbuffers::FlatBufferBuilder<'a, A>, root: flatbuffers::WIPOffset<WorkerAssignment<'a>>) {
+  fbb.finish_size_prefixed(root, None);
+}
