+++ crates/assignments/src/lib.rs
@@ -1,3 +1,39 @@
+// Raw flatc output, one module per schema file (schema/*.fbs) -- generated code cross-references
+// siblings via hardcoded `crate::<name>_generated::*` paths, so these must live at the crate
+// root. `assignment_fb` re-exports all of them into one flat namespace for internal use.
+#[allow(
+    dead_code,
+    unused_imports,
+    unsafe_op_in_unsafe_fn,
+    mismatched_lifetime_syntaxes,
+    clippy::all
+)]
+mod assignment_generated {
+    include!("../schema/gen/assignment_generated.rs");
+}
+#[cfg(feature = "mvcc-chunks")]
+#[allow(
+    dead_code,
+    unused_imports,
+    unsafe_op_in_unsafe_fn,
+    mismatched_lifetime_syntaxes,
+    clippy::all
+)]
+mod worker_assignment_generated {
+    include!("../schema/gen/worker_assignment_generated.rs");
+}
+#[cfg(feature = "mvcc-chunks")]
+#[allow(
+    dead_code,
+    unused_imports,
+    unsafe_op_in_unsafe_fn,
+    mismatched_lifetime_syntaxes,
+    clippy::all
+)]
+mod portal_assignment_generated {
+    include!("../schema/gen/portal_assignment_generated.rs");
+}
+
 mod assignment_fb;
 #[cfg(feature = "builder")]
 mod builder;
